# book-network-back/src/main/java/com/fakhri/book/feedback/FeedbackResponse.java
package com.fakhri.book.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}