# book-network-front/src/app/services/models/book-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookRequest {
  id?: number;
  authorName: string;
  isbn: string;
  shareable?: boolean;
  synopsis: string;
  title: string;
}